Refactor notification alert and iOS detection to avoid "NaN" infinite loop and improve home page flickering
